Edit: Update Week 01 Homework

+++ Grammar/c_operator.c
@@ -2,11 +2,11 @@
 #include <stdio.h>
 
 int main() {
-  int a = 1;
-  int b = 2;
-  printf("a + b 는 : %d\n", a + b);
-  printf("a - b 는 : %d\n", a - b);
-  printf("a * b 는 : %d\n", a * b);
-  printf("a / b 는 : %d\n", a / b);
+  float a = 1;
+  float b = 2;
+  printf("a + b 는 : %f\n", a + b);
+  printf("a - b 는 : %f\n", a - b);
+  printf("a * b 는 : %f\n", a * b);
+  printf("a / b 는 : %f\n", a / b);
   return 0;
 }
